% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestploter.R
\docType{package}
\name{forestploter}
\alias{forestploter}
\alias{forestploter-package}
\title{Create Forest Plot}
\description{
This package uses gtable and gridExtra to overlay forest plots.
}
\seealso{
\code{\link{grid}},\code{\link{gridExtra}}
}
\author{
Alimu Dayimu \email{alimdayim@hotmail.com}
}
\keyword{packagelibrary}
