% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_scale}
\alias{get_scale}
\title{Get the scale to fit the plot}
\usage{
get_scale(plot, width_wanted, height_wanted, unit = c("in", "cm", "mm"))
}
\arguments{
\item{plot}{A forest plot object.}

\item{width_wanted, height_wanted}{Desired plot size in \code{units} to save plot.}

\item{unit}{Unit of the plot size in \code{units} ("in", "cm", or "mm")to be saved.}
}
\value{
A numeric value
}
\description{
Use \code{get_scale} function to get the optimal scale with a wanted width and height to
print the plot. This function can be used with \code{\link[ggsave]{ggplot2}}
function.
}
\examples{
\dontrun{
 dt <- read.csv(system.file("extdata", "example_data.csv", package = "forestploter"))
 dt <- dt[1:6,1:6]

 dt$` ` <- paste(rep(" ", 20), collapse = " ")

 p <- forest(dt[,c(1:3, 7)],
             est = dt$est,
             lower = dt$low,
             upper = dt$hi,
             ci_column = 4)

# Use ggsave from ggplot2 package and scale
scale = get_scale(p, 5, 3)
ggplot2::ggsave("test.pdf", p,
       width = 5,
       height = 3,
       units = 'in',
       scale = scale)
}

}
