% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-arrow.R
\name{make_arrow}
\alias{make_arrow}
\title{Make arrow}
\usage{
make_arrow(x0 = 1, arrow_lab, arrow_gp, col_width, xlim, x_trans = "none")
}
\arguments{
\item{x0}{Position of vertical line for 0 or 1.}

\item{arrow_lab}{Labels for the arrows, a vector of length two.}

\item{arrow_gp}{Graphical parameters for arrow.}

\item{col_width}{Width of the column arrow to be fitted.}

\item{xlim}{Limits for the x-axis as a vector of length 2, i.e.
\code{c(low, high)}. By default the minimum and maximum of the lower and
upper values are used.}

\item{x_trans}{Scale of the axis, one of \code{"none"} (default),
\code{"log"}, \code{"log2"} or \code{"log10"}. Use \code{"log"} if the
values are exponential, e.g. odds ratios or hazard ratios. The default
reference line of \code{\link{forest}} is 1 for log scales and 0 otherwise.}
}
\description{
Make arrow
}
\keyword{internal}
