% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeci.R
\name{make_summary}
\alias{make_summary}
\title{Create pooled summary diamond shape}
\usage{
make_summary(est, lower, upper, sizes = 1, gp, xlim, nudge_y = 0)
}
\arguments{
\item{est}{Point estimation. Can be a list for multiple columns
and/or multiple groups. If the length of the list is larger than
then length of \code{ci_column}, then the values reused for each column
and considered as different groups.}

\item{lower}{Lower bound of the confidence interval, same as \code{est}.}

\item{upper}{Upper bound of the confidence interval, same as \code{est}.}

\item{sizes}{Size of the point estimation box, can be a vector or a list.
The value is a multiple of one line of text, so \code{1} draws a point as tall
as the \code{base_size} of the theme. The same scale applies to the summary
diamond. Values are used as they are, unless \code{\link{scale_sizes}} is
used to read them as study weights; useful values are roughly between
\code{0.2} and \code{1.5}, and a warning is given when the plot is drawn if
they are outside \code{0.1} to \code{2}.}

\item{gp}{Graphical parameters of \code{\link[grid]{gpar}}.
Please refer to \code{\link{forest_theme}} for more details.}

\item{xlim}{Limits for the x-axis as a vector of length 2, i.e.
\code{c(low, high)}. By default the minimum and maximum of the lower and
upper values are used.}

\item{nudge_y}{Vertical adjustment to nudge groups by, must be within 0 to 1.
Defaults to \code{0}; for grouped forest plots a value of \code{0} is bumped
to \code{0.1} automatically so that group CIs do not overplot. Set explicitly
to override.}
}
\value{
A gTree object
}
\description{
Create pooled summary diamond shape
}
