% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-xaixs.R
\name{make_xaxis}
\alias{make_xaxis}
\title{Create x-axis}
\usage{
make_xaxis(
  at,
  at_minor = NULL,
  xlab = NULL,
  x0 = 1,
  x_trans = "none",
  ticks_digits = 1,
  gp = gpar(),
  xlab_gp = NULL,
  xlim
)
}
\arguments{
\item{at}{Numerical vector, create ticks at given values.}

\item{at_minor}{Numerical vector, create ticks at given values without label.}

\item{xlab}{X-axis label, see \code{\link{set_labs}}.}

\item{x0}{Position of vertical line for 0 or 1.}

\item{x_trans}{Scale of the axis, one of \code{"none"} (default),
\code{"log"}, \code{"log2"} or \code{"log10"}. Use \code{"log"} if the
values are exponential, e.g. odds ratios or hazard ratios. The default
reference line of \code{\link{forest}} is 1 for log scales and 0 otherwise.}

\item{ticks_digits}{Number of digits for the tick labels. If an integer is
given, for example \code{1L}, trailing zeros after the decimal mark are
dropped. Give a double, for example \code{1}, to keep them. By default the
number is calculated from the tick positions. Use a list to mix the two
between CI columns, as a vector makes them all double.}

\item{gp}{Graphical parameters for arrow.}

\item{xlab_gp}{Graphical parameters for xlab.}

\item{xlim}{Limits for the x-axis as a vector of length 2, i.e.
\code{c(low, high)}. By default the minimum and maximum of the lower and
upper values are used.}
}
\value{
A grob
}
\description{
This function used to x-axis for the forest plot.
}
\keyword{internal}
