% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_xlim}
\alias{make_xlim}
\title{Create xlim}
\usage{
make_xlim(
  xlim = NULL,
  lower,
  upper,
  ref_line = ifelse(x_trans \%in\% c("log", "log2", "log10"), 1, 0),
  ticks_at = NULL,
  x_trans = "none"
)
}
\arguments{
\item{xlim}{Limits for the x-axis as a vector of length 2, i.e.
\code{c(low, high)}. By default the minimum and maximum of the lower and
upper values are used.}

\item{lower}{Lower bound of the confidence interval, same as \code{est}.}

\item{upper}{Upper bound of the confidence interval, same as \code{est}.}

\item{ref_line}{X-axis coordinates of the reference line, the value of no
effect. If \code{NULL} (default), it is 1 if the x-axis is on a log scale
(see \code{\link{set_xaxis}}) and 0 otherwise. Provide an atomic vector if
different reference line for each \code{ci_column} is desired.}

\item{ticks_at}{Tick mark positions. By default, ticks are computed
automatically: \code{\link[base]{pretty}} for linear axes and a decade-aware
helper for log scales (e.g. \code{0.1, 1, 10, 100} for a wide log10 range).}

\item{x_trans}{Scale of the axis, one of \code{"none"} (default),
\code{"log"}, \code{"log2"} or \code{"log10"}. Use \code{"log"} if the
values are exponential, e.g. odds ratios or hazard ratios. The default
reference line of \code{\link{forest}} is 1 for log scales and 0 otherwise.}
}
\value{
A list
}
\description{
Create xlim based on value ranges.
}
\keyword{internal}
