% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makexaxis.R
\name{makexaxis}
\alias{makexaxis}
\title{Make x-axis and arrow}
\usage{
makexaxis(at, x0 = 1, labels = NULL, xlim, arrow.lab = NULL)
}
\arguments{
\item{at}{A numeric vector of x-value locations for the tick marks.}

\item{x0}{Position of vertical line for 0 or 1.}

\item{labels}{Labels of the tick}

\item{xlim}{Limits for the x axis as a vector length 2, i.e. c(low, high)}

\item{arrow.lab}{Label for the arrow, left and right.}
}
\description{
Make x-axis and arrow
}
