% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{print.forest_style}
\alias{print.forest_style}
\title{Print a forest plot style}
\usage{
\method{print}{forest_style}(x, ...)
}
\arguments{
\item{x}{A style created with \code{\link{forest_style}}.}

\item{...}{other arguments not used by this method}
}
\value{
Invisibly returns the style.
}
\description{
Show the settings of a style that are not the default ones.
}
