% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{print.forestplot}
\alias{print.forestplot}
\alias{plot.forestplot}
\title{Draw plot}
\usage{
\method{print}{forestplot}(x, autofit = FALSE, ...)

\method{plot}{forestplot}(x, autofit = FALSE, ...)
}
\arguments{
\item{x}{forestplot to display}

\item{autofit}{If true, the page is shared equally between the columns and
between the rows of the plot. This will be deprecated, use \code{fit} of
\code{\link{forest_style}} instead, which also works with
\code{ggplot2::ggsave} and \code{patchwork}.}

\item{...}{other arguments not used by this method}
}
\value{
Invisibly returns the original forestplot.
}
\description{
Print or draw forestplot.
}
