% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_grob.R
\name{set_legend}
\alias{set_legend}
\title{CI settings for group}
\usage{
set_legend(name = "", position = NULL, value = NULL, pch = 15, color)
}
\arguments{
\item{name}{Title of the legend.}

\item{position}{Position of the legend, \code{"right"}, \code{"top"},
\code{"bottom"}.}

\item{value}{Legend labels (expressions)}

\item{pch}{Shape of the point symbol, see \code{\link[grid]{pointsGrob}}.}

\item{color}{Colors for the group/confidence interval.}
}
\description{
This is to set the point group name, symbol and color of the CI.
}
