% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{vert_line}
\alias{vert_line}
\title{Add vertical line}
\usage{
vert_line(x, gp = grid::gpar(), xlim)
}
\arguments{
\item{x}{A numeric unit object specifying x-values.}

\item{gp}{An object of class \code{gpar}, see \code{\link[grid]{gpar}}.}

\item{xlim}{Limits for the x axis as a vector length 2, i.e. c(low, high)}
}
\description{
Add vertical line
}
