% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_errors.r
\name{check_errors}
\alias{check_errors}
\title{Checking error for forest plot}
\usage{
check_errors(data, est, lower, upper, sizes, ref_line, ci_column, is_summary)
}
\arguments{
\item{data}{Data to be displayed in the forest plot}

\item{est}{Point estimation. Can be a list for multiple columns
and/or multiple groups. If the length of the list is larger than
then length of \code{ci_column}, then the values reused for each column
and considered as different groups.}

\item{lower}{Lower bound of the confidence interval, same as \code{est}.}

\item{upper}{Upper bound of the confidence interval, same as \code{est}.}

\item{sizes}{Size of the point estimation box, can be a vector or a list.
The value is a multiple of one line of text, so \code{1} draws a point as tall
as the \code{base_size} of the theme. The same scale applies to the summary
diamond. Values are used as they are, unless \code{\link{scale_sizes}} is
used to read them as study weights; useful values are roughly between
\code{0.2} and \code{1.5}, and a warning is given when the plot is drawn if
they are outside \code{0.1} to \code{2}.}

\item{ref_line}{X-axis coordinates of the reference line, the value of no
effect. If \code{NULL} (default), it is 1 if the x-axis is on a log scale
(see \code{\link{set_xaxis}}) and 0 otherwise. Provide an atomic vector if
different reference line for each \code{ci_column} is desired.}

\item{ci_column}{Column number of the data the CI will be displayed.}

\item{is_summary}{A logical vector indicating if the value is a summary value,
which will have a diamond shape for the estimate. With multiple groups the
diamonds are stacked in the same cell and the summary rows are made taller to
fit them, so a larger \code{nudge_y} may be wanted.}
}
\description{
The settings of the x-axis and the labels are checked by
\code{\link{set_xaxis}} and \code{\link{set_labs}}.
}
\keyword{internal}
