% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_plot.R
\name{edit_plot}
\alias{edit_plot}
\title{Edit forest plot}
\usage{
edit_plot(
  plot,
  row = NULL,
  col = NULL,
  part = c("body", "header"),
  which = c("text", "background"),
  gp
)
}
\arguments{
\item{plot}{A forest plot object.}

\item{row}{A numeric value or vector indicating row number to edit in the
dataset. Will edit the whole row if left blank for the body. This will be
ignored if the \code{part} is "header".}

\item{col}{A numeric value or vector indicating column to edit in the dataset.
Will edit the whole column if left blank.}

\item{part}{Part to edit, body (default) or header.}

\item{which}{Which element to edit, text or background of the cell.}

\item{gp}{Pass \code{gpar} parameters, see \code{\link[grid]{gpar}}. It should
be passed as \code{gpar(col = "red")}.}
}
\value{
A \code{\link[gtable]{gtable}} object.
}
\description{
This function is used to edit the graphical parameter of the forest plot.
}
