% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{forest}
\alias{forest}
\title{Forest plot}
\usage{
forest(
  data,
  est,
  lower,
  upper,
  sizes = 0.4,
  null.at = 1,
  ci.col,
  ci.width = 1,
  xlim = NULL,
  tick.breaks = NULL,
  arrow.lab = NULL,
  theme = NULL
)
}
\arguments{
\item{data}{Data to be displayed in the forest plot}

\item{est}{Point estimation}

\item{lower}{Lower bound of the confidence interval}

\item{upper}{Upper bound of the confidence interval}

\item{sizes}{Size of the point estimation box.}

\item{null.at}{X-axis coordinates of zero line, default is 1.}

\item{ci.col}{Column number of the data the CI will be displayed.}

\item{ci.width}{Width of the CI column in the forest plot, default is 1. This
is the ratio of maximum width of the columns in the data.}

\item{xlim}{Limits for the x axis as a vector length 2, i.e. c(low, high). It
will take the maximum and minimum of the \code{tick.breaks} or CI.}

\item{tick.breaks}{X-axis breaks points, a vector.}

\item{arrow.lab}{Labels for the arrows, string vector of length two (left and
right).}

\item{theme}{Theme of the forest plot, see \link\link[gridExtra:tableGrob]{gridExtra::tableGrob} for
details.}
}
\value{
A gtable object.
}
\description{
Forest plot
}
