% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{forest}
\alias{forest}
\title{Forest plot}
\usage{
forest(
  data,
  est,
  lower,
  upper,
  sizes = 0.4,
  ref_line = 1,
  ci_column,
  xlim = NULL,
  xaxis = NULL,
  arrow_lab = NULL,
  footnote = NULL,
  nudge_y = 0,
  theme = NULL
)
}
\arguments{
\item{data}{Data to be displayed in the forest plot}

\item{est}{Point estimation. Can be a list for multiple columns
and/or multiple groups. If the length of the list is larger than
then length of \code{ci_column}, then the values reused for each column
and considered as different groups.}

\item{lower}{Lower bound of the confidence interval, same as \code{est}.}

\item{upper}{Upper bound of the confidence interval, same as \code{est}.}

\item{sizes}{Size of the point estimation box, can be a unit, vector or a list.}

\item{ref_line}{X-axis coordinates of zero line, default is 1.}

\item{ci_column}{Column number of the data the CI will be displayed.}

\item{xlim}{Limits for the x axis as a vector of length 2, i.e. c(low, high). It
will take the minimum and maximum of the lower and upper value if not provided.}

\item{xaxis}{Set X-axis breaks points and labels, should use \code{\link{set_xaxis}}.}

\item{arrow_lab}{Labels for the arrows, string vector of length two (left and
right). The theme of arrow will inherit from the x-axis.}

\item{footnote}{Footnote for the forest plot, will be aligned at left bottom
of the plot. Please adjust the line length with line break to avoid the overlap
with the arrow and/or x-axis.}

\item{nudge_y}{Horizontal adjustment to nudge groups by, must be within 0 to 1.}

\item{theme}{Theme of the forest plot, see \code{\link{forest_theme}} for
details.}
}
\value{
A gtable object.
}
\description{
A data frame will be used for the basic layout of the forest plot.
Graphical parameters can be set using the \code{\link{forest_theme}} function.
}
