% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{forest}
\alias{forest}
\title{Forest plot}
\usage{
forest(
  data,
  est,
  lower,
  upper,
  sizes = 0.4,
  ref_line = ifelse(xlog, 1, 0),
  vert_line = NULL,
  ci_column,
  xlog = FALSE,
  is_summary = NULL,
  xlim = NULL,
  ticks_at = NULL,
  arrow_lab = NULL,
  footnote = NULL,
  nudge_y = 0,
  theme = NULL
)
}
\arguments{
\item{data}{Data to be displayed in the forest plot}

\item{est}{Point estimation. Can be a list for multiple columns
and/or multiple groups. If the length of the list is larger than
then length of \code{ci_column}, then the values reused for each column
and considered as different groups.}

\item{lower}{Lower bound of the confidence interval, same as \code{est}.}

\item{upper}{Upper bound of the confidence interval, same as \code{est}.}

\item{sizes}{Size of the point estimation box, can be a unit, vector or a list.}

\item{ref_line}{X-axis coordinates of zero line, default is 1.}

\item{vert_line}{Numerical vector, add additional vertical line at given value.}

\item{ci_column}{Column number of the data the CI will be displayed.}

\item{xlog}{If TRUE, x-axis tick marks assume values are exponential, e.g.
for logistic regression (OR), survival estimates (HR), Poisson regression etc.}

\item{is_summary}{A logical vector indicating if the value is a summary value,
which will have a diamond shape for the estimate. Can not be used with multiple
group \code{forestplot}.}

\item{xlim}{Limits for the x axis as a vector of length 2, i.e. c(low, high). It
will take the minimum and maximum of the lower and upper value if not provided.
This will apply to all CI columns if provided, and will be calculated automatically
each column if not provided.}

\item{ticks_at}{Set X-axis tick-marks point. This will apply to all CI columns if
provided, and will be calculated automatically for each column if not provided.}

\item{arrow_lab}{Labels for the arrows, string vector of length two (left and
right). The theme of arrow will inherit from the x-axis.}

\item{footnote}{Footnote for the forest plot, will be aligned at left bottom
of the plot. Please adjust the line length with line break to avoid the overlap
with the arrow and/or x-axis.}

\item{nudge_y}{Horizontal adjustment to nudge groups by, must be within 0 to 1.}

\item{theme}{Theme of the forest plot, see \code{\link{forest_theme}} for
details.}
}
\value{
A \code{\link[gtable]{gtable}} object.
}
\description{
A data frame will be used for the basic layout of the forest plot.
Graphical parameters can be set using the \code{\link{forest_theme}} function.
}
\examples{
# Read provided sample example data
dt <- read.csv(system.file("extdata", "example_data.csv", package = "forestploter"))

# Keep needed columns
dt <- dt[,1:6]

# indent the subgroup if there is a number in the placebo column
dt$Subgroup <- ifelse(is.na(dt$Placebo),
                      dt$Subgroup,
                      paste0("   ", dt$Subgroup))

# NA to blank or NA will be transformed to carachter.
dt$Treatment <- ifelse(is.na(dt$Treatment), "", dt$Treatment)
dt$Placebo <- ifelse(is.na(dt$Placebo), "", dt$Placebo)
dt$se <- (log(dt$hi) - log(dt$est))/1.96

# Add blank column for the forest plot to display CI.
# Adjust the column width with space.
dt$` ` <- paste(rep(" ", 20), collapse = " ")

# Create confidence interval column to display
dt$`HR (95\% CI)` <- ifelse(is.na(dt$se), "",
                             sprintf("\%.2f (\%.2f to \%.2f)",
                                     dt$est, dt$low, dt$hi))

# Define theme
tm <- forest_theme(base_size = 10,
                   refline_col = "red",
                   footnote_col = "#636363",
                   footnote_fontface = "italic")

p <- forest(dt[,c(1:3, 8:9)],
            est = dt$est,
            lower = dt$low,
            upper = dt$hi,
            sizes = dt$se,
            ci_column = 4,
            ref_line = 1,
            arrow_lab = c("Placebo Better", "Treatment Better"),
            xlim = c(0, 4),
            ticks_at = c(0.5, 1, 2, 3),
            footnote = "This is the demo data. Please feel free to change\nanything you want.",
            theme = tm)

# Print plot
plot(p)
}
