% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{forest_theme}
\alias{forest_theme}
\title{Forest plot default theme}
\usage{
forest_theme(
  base_size = 12,
  base_family = "",
  xaxis_lwd = 0.6,
  refline_lwd = 1,
  refline_lty = "dashed",
  refline_col = "grey20",
  ...
)
}
\arguments{
\item{base_size}{The size of text}

\item{base_family}{The font family}

\item{xaxis_lwd}{Line width for x-axis.}

\item{refline_lwd}{Line width for reference line.}

\item{refline_lty}{Line type for reference line.}

\item{refline_col}{Line color for the reference line.}

\item{...}{Other parameters passed to table. See \code{\link[gridExtra]{tableGrob}}
for details.}
}
\description{
Default theme for the forest plot, but can pass other parameters. The
parameters will be passed to corresponding elements of the forest plot.
See \code{\link[grid]{gpar}} for details.
}
