% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestploter.R
\docType{package}
\name{forestploter-package}
\alias{forestploter}
\alias{forestploter-package}
\title{forestploter: Create a Flexible Forest Plot}
\description{
Create a forest plot based on the layout of the data. Confidence intervals in multiple columns by groups can be done easily. Editing the plot, inserting/adding text, applying a theme to the plot, and much more.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/adayim/forestploter}
  \item Report bugs at \url{https://github.com/adayim/forestploter/issues}
}

}
\author{
\strong{Maintainer}: Alimu Dayimu \email{ad938@cam.ac.uk} (\href{https://orcid.org/0000-0001-9998-7463}{ORCID})

}
\keyword{internal}
