% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_grob.R
\name{legend_grob}
\alias{legend_grob}
\title{Create legends}
\usage{
legend_grob(
  name = "",
  label,
  position = c("right", "top", "bottom"),
  hgap = unit(0.1, "lines"),
  vgap = unit(0.5, "lines"),
  pch = 15,
  ncol = 1,
  gp = gpar(lty = 1, col = "black", fill = "black", fontsize = 12, fontfamily = ""),
  byrow = TRUE,
  ...
)
}
\arguments{
\item{name}{Character string, Legend name.}

\item{label}{legend labels (expressions).}

\item{position}{Position of the legend, \code{"right"}, \code{"top"},
\code{"bottom"}.}

\item{hgap}{Horizontal gap between the legend entries,
see \code{\link[grid]{legendGrob}} for details.}

\item{vgap}{Vertical gap between the legend entries,
see \code{\link[grid]{legendGrob}} for details.}

\item{pch}{Legend symbol.}

\item{ncol}{integer; the number of columns}

\item{gp}{Graphical parameters.}

\item{byrow}{logical indicating whether rows of the legend are filled first.}

\item{...}{Other parameters, not used currently.}
}
\value{
A frame grob
}
\description{
This function used to create legends for the forest plot.
}
\keyword{internal}
