% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend_grob.R
\name{legend_grob}
\alias{legend_grob}
\title{Create legends}
\usage{
legend_grob(
  name = "",
  label,
  color,
  position = c("right", "top", "bottom"),
  hgap = unit(0.1, "lines"),
  vgap = unit(0.5, "lines"),
  pch = 15,
  lty = 1,
  fontsize = 12,
  fontfamily = "",
  ...
)
}
\arguments{
\item{name}{Character string, Legend name.}

\item{label}{legend labels (expressions).}

\item{color}{Colors for the group}

\item{position}{Position of the legend, \code{"right"}, \code{"top"},
\code{"bottom"}.}

\item{hgap}{Horizontal gap between the legend entries,
see \code{\link[grid]{legendGrob}} for details.}

\item{vgap}{Vertical gap between the legend entries,
see \code{\link[grid]{legendGrob}} for details.}

\item{pch}{Legend symbol.}

\item{lty}{Line type.}

\item{fontsize}{Font size of the legend.}

\item{fontfamily}{Font family of the legend.}

\item{...}{Other parameters, not used currently.}
}
\value{
A frame grob
}
\description{
This function used to create legends for the forest plot.
}
\keyword{internal}
