% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-arrow.R
\name{make_arrow}
\alias{make_arrow}
\title{Make arrow}
\usage{
make_arrow(x0 = 1, arrow_lab, gp, xlim, is_exp = FALSE)
}
\arguments{
\item{x0}{Position of vertical line for 0 or 1.}

\item{arrow_lab}{Label for the arrow, left and right.}

\item{gp}{Graphical parameters for arrow.}

\item{xlim}{Limits for the x axis as a vector length 2, i.e. c(low, high)}

\item{is_exp}{If values is exponential.}
}
\description{
Make arrow
}
\keyword{internal}
