% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_ticks}
\alias{make_ticks}
\title{Set x-axis ticks}
\usage{
make_ticks(at = NULL, xlim, refline = ifelse(is_exp, 1, 0), is_exp = FALSE)
}
\arguments{
\item{at}{Numerical vector, create ticks at given values.}

\item{xlim}{Limits for the x axis as a vector of length 2, i.e. c(low, high). It
will take the minimum and maximum of the lower and upper value if not provided.
This will apply to all CI columns if provided, and will be calculated automatically
each column if not provided.}

\item{is_exp}{If values is exponential.}
}
\value{
A vector
}
\description{
Create ticks points.
}
\keyword{internal}
