% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-xaixs.R
\name{make_xaxis}
\alias{make_xaxis}
\title{Create xaxis}
\usage{
make_xaxis(
  at,
  xlab = NULL,
  x0 = 1,
  is_exp = FALSE,
  ticks_digits = 1,
  gp = gpar(),
  xlim
)
}
\arguments{
\item{at}{Numerical vector, create ticks at given values.}

\item{xlab}{X-axis label.}

\item{x0}{Position of vertical line for 0 or 1.}

\item{is_exp}{If values is exponential.}

\item{ticks_digits}{Number of digits for the x-axis, default is 1.}

\item{gp}{Graphical parameters for arrow.}

\item{xlim}{Limits for the x axis as a vector length 2, i.e. c(low, high)}
}
\value{
A grob
}
\description{
This function used to xais for the forest plot.
}
\keyword{internal}
