% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-xaixs.R
\name{make_xaxis}
\alias{make_xaxis}
\title{Create xaxis}
\usage{
make_xaxis(at, xlab = NULL, is_exp = FALSE, gp = gpar(), xlim)
}
\arguments{
\item{at}{Numerical vector, create ticks at given values.}

\item{xlab}{X-axis label.}

\item{is_exp}{If values is exponential.}

\item{gp}{Graphical parameters for arrow.}

\item{xlim}{Limits for the x axis as a vector length 2, i.e. c(low, high)}
}
\value{
A grob
}
\description{
This function used to xais for the forest plot.
}
\keyword{internal}
