% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeci.R
\name{makeci}
\alias{makeci}
\title{Create confidence interval}
\usage{
makeci(est, lower, upper, size = 1, xlim = c(0, 1))
}
\arguments{
\item{est}{Point estimates, numeric}

\item{lower}{Lower bound}

\item{upper}{Upper bound}

\item{size}{Size of the point}

\item{xlim}{Limits for the x axis as a vector length 2, i.e. c(low, high)}
}
\description{
Create confidence interval
}
