% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeci.R
\name{makeci}
\alias{makeci}
\title{Create confidence interval}
\usage{
makeci(
  est,
  lower,
  upper,
  pch,
  size = 1,
  gp = gpar(),
  t_height = NULL,
  xlim = c(0, 1),
  nudge_y = 0
)
}
\arguments{
\item{est}{Point estimates, numeric}

\item{lower}{Lower bound}

\item{upper}{Upper bound}

\item{pch}{Numeric or character vector indicating what sort of plotting
symbol to use. See \code{\link[grid]{pointsGrob}}.}

\item{size}{Size of the point}

\item{gp}{Grphical parameters.}

\item{t_height}{The height confidence interval line end vertices. If
value is \code{NULL} (default), no vertices will be drawn.}

\item{xlim}{Limits for the x axis as a vector length 2, i.e. c(low, high)}

\item{nudge_y}{Offset Y coordinates.}

\item{color}{Color of the point and the line}
}
\description{
Create confidence interval
}
\keyword{internal}
