% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_xaxis}
\alias{set_xaxis}
\title{Set x-axis}
\usage{
set_xaxis(break_at, label_at = NULL, label_value = NULL)
}
\arguments{
\item{break_at}{Numerical vector, create ticks at given values.}

\item{label_at}{Optional, numerical vector. Put labels at given values. The
value of \code{break_at} will be used if not provided.}

\item{label_value}{Optional, character or numeric vector to put labels for the
the x-axis. Should have same length as the \code{label_at}. The value of the
value of \code{label_at} will be used if not provided.}
}
\value{
A list.
}
\description{
Set x-axis
}
