% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.r
\name{xscale}
\alias{xscale}
\title{Change Axis Scale: log2, log10 and more}
\usage{
xscale(
  x,
  scale = c("none", "log", "log2", "log10", "scientific"),
  type = c("scale", "inv", "format"),
  format_digits = 1
)
}
\arguments{
\item{scale}{axis scale. Allowed values are one of c("none", "log2", "log10",
"sqrt", "percent", "dollar", "scientific"); e.g.: .scale="log2".}

\item{format_digits}{Digits to keep while formating}

\item{inv}{Inverse value back to its orignal.}

\item{format}{logical value. If TRUE, axis tick mark labels will be formatted
when scale  = "log2" or "log10".}
}
\description{
Change axis scale.
}
\keyword{internal}
